Address title coalescing guard coverage
